# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/students": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "getStudentList",
                "description": "Get all Students",
                "operationId": "e694fbee0861f8e89913a14f6cf085bd",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Student"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "createStudent",
                "description": "Create Student",
                "operationId": "326e20b7fbcb950422a8f4c0079cd805",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Student"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Student"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/students/{id}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "getStudentItem",
                "description": "Get Student",
                "operationId": "d4d65b8b5ab661594ec4db4e2f798694",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Student",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Student"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Student"
                ],
                "summary": "updateStudent",
                "description": "Update Student",
                "operationId": "fc3cf67ef8dbab103fa5a4b454e5caf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Student",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Student"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Student"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Student"
                ],
                "summary": "deleteStudent",
                "description": "Delete Student",
                "operationId": "7379476f353883cd8a2e1aa428120c8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Student",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Student": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "gender": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "date_of_birth": {
                        "description": "",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    }
}